+++ src/articles/components/ArticleView/ArticleHeader.tsx
@@ -1,6 +1,6 @@
-import { DateTime } from 'luxon';
 import React, { FC } from 'react';
-import { IArticle } from '../../models/Article';
+
+import { IArticle } from 'articles/models/Article';
 import style from './articleView.less';
 
 export interface IProps {
@@ -8,27 +8,13 @@ export interface IProps {
 }
 
 export const ArticleHeader: FC<IProps> = ({ article }) => {
-  const { heading, ingress, published_date, changed_date, authors } = article;
-  const pubDateTime = DateTime.fromISO(published_date);
-  const changeDateTime = DateTime.fromISO(changed_date);
+  const { heading, ingress } = article;
 
   return (
     <header className={style.articleHeader}>
       <h1>{heading}</h1>
       <hr />
       <p>{ingress}</p>
-      <div className={style.byline}>
-        <span>Publisert </span>
-        <time dateTime={published_date}>{pubDateTime.toLocaleString()}</time>
-        <span> Skrevet av </span>
-        {authors}
-        {published_date !== changed_date && pubDateTime < changeDateTime ? (
-          <>
-            <span className={style.lastChanged}>Sist endret </span>
-            <time>{changeDateTime.toLocaleString()}</time>
-          </>
-        ) : null}
-      </div>
     </header>
   );
 };
